tests with stub part 1

+++ WorkshopProject/System Service/IGodObject.cs
@@ -184,6 +184,7 @@ namespace WorkshopProject.System_Service
         //disabled for now
         public void cleanUpAllData()
         {
+            clearDb();
             //new Repo().Delete();
 
             //WorkShop.stores = new Dictionary<int, Store>();
@@ -202,8 +203,16 @@ namespace WorkshopProject.System_Service
 
         public void clearDb()
         {
-            DataAccessDriver.clearStub();
-            new Repo().Delete();
+            if (DataAccessDriver.UseStub)
+            {
+                DataAccessDriver.clearStub();
+            }
+
+            else
+            {
+                new Repo().Delete();    
+            }
+            
         }
     }
 }
